// client/src/types/type.d.ts
export type ITestimonialsData = {
    id: number;
    text: string;
    author: string;
    image: string;
    jobTitle: string;
}


export type IPlans = {
    id: string;
    price: number;
    credits: number;
    desc: string;
}

export type IStepBox = {
    icon: string;
    title : string;
    description : string;
}